chore(server/binding/tikv): split kv shared parts
